+++ pinet-rest/src/main/java/com/pinet/rest/entity/dto/CartListDto.java
@@ -20,7 +20,7 @@ public class CartListDto {
 
     @ApiModelProperty(value = "店铺id",name = "shopId")
     @NotNull(message = "店铺id不能为空")
-    private Integer shopId;
+    private Long shopId;
 
 
     @ApiModelProperty(value = "用户id(内部封装参数，不需要传)",name = "customerId")
